structurial changes for rosbag2
