+++ rosbag2/src/rosbag2/writer.cpp
@@ -23,199 +23,41 @@
 
 #include "rosbag2/info.hpp"
 #include "rosbag2/storage_options.hpp"
+#include "rosbag2/writer_interfaces/base_writer_interface.hpp"
+
 #include "rosbag2_storage/filesystem_helper.hpp"
 
 namespace rosbag2
 {
 
-namespace
-{
-std::string format_storage_uri(const std::string & base_folder, uint64_t storage_count)
-{
-  // Right now `base_folder_` is always just the folder name for where to install the bagfile.
-  // The name of the folder needs to be queried in case Writer is opened with a relative path.
-  std::stringstream storage_file_name;
-  storage_file_name << rosbag2_storage::FilesystemHelper::get_folder_name(base_folder) <<
-    "_" << storage_count;
-
-  return rosbag2_storage::FilesystemHelper::concat({base_folder, storage_file_name.str()});
-}
-}  // namespace
-
-Writer::Writer(
-  std::unique_ptr<rosbag2_storage::StorageFactoryInterface> storage_factory,
-  std::shared_ptr<SerializationFormatConverterFactoryInterface> converter_factory,
-  std::unique_ptr<rosbag2_storage::MetadataIo> metadata_io)
-: storage_factory_(std::move(storage_factory)),
-  converter_factory_(std::move(converter_factory)),
-  storage_(nullptr),
-  metadata_io_(std::move(metadata_io)),
-  converter_(nullptr),
-  max_bagfile_size_(rosbag2_storage::storage_interfaces::MAX_BAGFILE_SIZE_NO_SPLIT),
-  topics_names_to_info_(),
-  metadata_()
+Writer::Writer(std::unique_ptr<rosbag2::writer_interfaces::BaseWriterInterface> writer_impl)
+: writer_impl_(std::move(writer_impl))
 {}
 
 Writer::~Writer()
 {
-  if (!base_folder_.empty()) {
-    finalize_metadata();
-    metadata_io_->write_metadata(base_folder_, metadata_);
-  }
-
-  storage_.reset();  // Necessary to ensure that the storage is destroyed before the factory
-  storage_factory_.reset();
-}
-
-void Writer::init_metadata()
-{
-  metadata_ = rosbag2_storage::BagMetadata{};
-  metadata_.storage_identifier = storage_->get_storage_identifier();
-  metadata_.starting_time = std::chrono::time_point<std::chrono::high_resolution_clock>(
-    std::chrono::nanoseconds::max());
-  metadata_.relative_file_paths = {storage_->get_relative_file_path()};
+  writer_impl_.reset();
 }
 
 void Writer::open(
-  const StorageOptions & storage_options,
-  const ConverterOptions & converter_options)
+  const StorageOptions & storage_options, const ConverterOptions & converter_options)
 {
-  max_bagfile_size_ = storage_options.max_bagfile_size;
-  base_folder_ = storage_options.uri;
-
-  if (converter_options.output_serialization_format !=
-    converter_options.input_serialization_format)
-  {
-    converter_ = std::make_unique<Converter>(converter_options, converter_factory_);
-  }
-
-  const auto storage_uri = format_storage_uri(base_folder_, 0);
-
-  storage_ = storage_factory_->open_read_write(storage_uri, storage_options.storage_id);
-  if (!storage_) {
-    throw std::runtime_error("No storage could be initialized. Abort");
-  }
-
-  init_metadata();
+  writer_impl_->open(storage_options, converter_options);
 }
 
 void Writer::create_topic(const TopicMetadata & topic_with_type)
 {
-  if (!storage_) {
-    throw std::runtime_error("Bag is not open. Call open() before writing.");
-  }
-
-  if (converter_) {
-    converter_->add_topic(topic_with_type.name, topic_with_type.type);
-  }
-
-  if (topics_names_to_info_.find(topic_with_type.name) ==
-    topics_names_to_info_.end())
-  {
-    rosbag2_storage::TopicInformation info{};
-    info.topic_metadata = topic_with_type;
-
-    const auto insert_res = topics_names_to_info_.insert(
-      std::make_pair(topic_with_type.name, info));
-
-    if (!insert_res.second) {
-      std::stringstream errmsg;
-      errmsg << "Failed to insert topic \"" << topic_with_type.name << "\"!";
-
-      throw std::runtime_error(errmsg.str());
-    }
-
-    storage_->create_topic(topic_with_type);
-  }
+  writer_impl_->create_topic(topic_with_type);
 }
 
 void Writer::remove_topic(const TopicMetadata & topic_with_type)
 {
-  if (!storage_) {
-    throw std::runtime_error("Bag is not open. Call open() before removing.");
-  }
-
-  if (topics_names_to_info_.erase(topic_with_type.name) > 0) {
-    storage_->remove_topic(topic_with_type);
-  } else {
-    std::stringstream errmsg;
-    errmsg << "Failed to remove the non-existing topic \"" <<
-      topic_with_type.name << "\"!";
-
-    throw std::runtime_error(errmsg.str());
-  }
-}
-
-void Writer::split_bagfile()
-{
-  const auto storage_uri = format_storage_uri(
-    base_folder_,
-    metadata_.relative_file_paths.size());
-  storage_ = storage_factory_->open_read_write(storage_uri, metadata_.storage_identifier);
-
-  if (!storage_) {
-    std::stringstream errmsg;
-    errmsg << "Failed to rollover bagfile to new file: \"" << storage_uri << "\"!";
-
-    throw std::runtime_error(errmsg.str());
-  }
-
-  metadata_.relative_file_paths.push_back(storage_->get_relative_file_path());
-
-  // Re-register all topics since we rolled-over to a new bagfile.
-  for (const auto & topic : topics_names_to_info_) {
-    storage_->create_topic(topic.second.topic_metadata);
-  }
+  writer_impl_->remove_topic(topic_with_type);
 }
 
 void Writer::write(std::shared_ptr<SerializedBagMessage> message)
 {
-  if (!storage_) {
-    throw std::runtime_error("Bag is not open. Call open() before writing.");
-  }
-
-  // Update the message count for the Topic.
-  ++topics_names_to_info_.at(message->topic_name).message_count;
-
-  if (should_split_bagfile()) {
-    split_bagfile();
-  }
-
-  const auto message_timestamp = std::chrono::time_point<std::chrono::high_resolution_clock>(
-    std::chrono::nanoseconds(message->time_stamp));
-  metadata_.starting_time = std::min(metadata_.starting_time, message_timestamp);
-
-  const auto duration = message_timestamp - metadata_.starting_time;
-  metadata_.duration = std::max(metadata_.duration, duration);
-
-  storage_->write(converter_ ? converter_->convert(message) : message);
-}
-
-bool Writer::should_split_bagfile() const
-{
-  if (max_bagfile_size_ == rosbag2_storage::storage_interfaces::MAX_BAGFILE_SIZE_NO_SPLIT) {
-    return false;
-  } else {
-    return storage_->get_bagfile_size() > max_bagfile_size_;
-  }
-}
-
-void Writer::finalize_metadata()
-{
-  metadata_.bag_size = 0;
-
-  for (const auto & path : metadata_.relative_file_paths) {
-    metadata_.bag_size += rosbag2_storage::FilesystemHelper::get_file_size(path);
-  }
-
-  metadata_.topics_with_message_count.clear();
-  metadata_.topics_with_message_count.reserve(topics_names_to_info_.size());
-  metadata_.message_count = 0;
-
-  for (const auto & topic : topics_names_to_info_) {
-    metadata_.topics_with_message_count.push_back(topic.second);
-    metadata_.message_count += topic.second.message_count;
-  }
+  writer_impl_->write(message);
 }
 
 }  // namespace rosbag2
